src/patient-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
azure-servicebus==7.12.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
psycopg2-binary==2.9.9
redis==5.0.1
pytest==7.4.4
pytest-asyncio==0.23.3
